renamed calendar file to capital case; added loading animation for dots when calendar loading; added indicators for hover on events; changed around hover to be on individual event, rather than cell
